pyproject.toml
[project]
name = "power-plant-rl"
version = "0.1.0"
description = "Reinforcement learning project for power plant optimization"
authors = [
    {name = "BgDaniel",email = "daniel.bergermat@t-online.de"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "numpy (>=2.3.2,<3.0.0)",
    "pandas (>=2.3.2,<3.0.0)",
    "xarray (>=2025.9.0,<2026.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "pathlib (>=1.0.1,<2.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "numba (>=0.62.0,<0.63.0)",
    "torch (>=2.10.0,<3.0.0)"
]

[tool.poetry]

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
black = "^25.1.0"
flake8 = "^7.3.0"
mypy = "^1.17.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
